+++ src/pages/admin/comprobantes.tsx
@@ -41,6 +41,7 @@ const columns:GridColDef[] = [
             field: 'check',
             headerName: 'Acción',
             renderCell: ({ row }: GridValueGetterParams | GridRenderCellParams ) => {
+                if ((user?.role !== 'admin') && (user?._id !== row.userInv)) return null;
                 return (
                         <Chip variant='outlined' label="Eliminar" color="error"
                         onClick={() => deleteHandler(row.id)}
